# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: APK Fraud Intelligence Investigator API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /investigations:
    get:
      operationId: listInvestigations
      tags: [investigations]
      summary: List investigations
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: riskLevel
          in: query
          required: false
          schema:
            type: string
            enum: [Low, Medium, High, Critical]
      responses:
        "200":
          description: List of investigations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InvestigationSummary"
    post:
      operationId: createInvestigation
      tags: [investigations]
      summary: Submit APK forensic data for autonomous investigation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvestigationInput"
      responses:
        "201":
          description: Investigation created and analyzed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Investigation"
  /investigations/{id}:
    get:
      operationId: getInvestigation
      tags: [investigations]
      summary: Get full investigation details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Investigation details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Investigation"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteInvestigation
      tags: [investigations]
      summary: Delete an investigation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /investigations/{id}/iocs:
    get:
      operationId: getInvestigationIocs
      tags: [investigations]
      summary: Get IOCs for a single investigation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: IOC list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ioc"
...
  /campaigns/{clusterId}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Campaign detail
      parameters:
        - name: clusterId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Campaign with members
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    CreateInvestigationInput:
      type: object
      properties:
        sampleName:
          type: string
          description: Display name for this APK sample (e.g. com.bank.fake-1.2.apk)
        sha256:
          type: string
        fuzzyHash:
          type: ["string", "null"]
        packageName:
          type: ["string", "null"]
        versionName:
          type: ["string", "null"]
        targetSdk:
          type: ["integer", "null"]
        compileSdk:
          type: ["integer", "null"]
        permissions:
          type: array
          items:
            type: string
        codeSnippets:
          type: ["string", "null"]
          description: Decompiled Java/Kotlin snippets
        urls:
          type: array
          items:
            type: string
        domains:
          type: array
          items:
            type: string
        ipAddresses:
          type: array
          items:
            type: string
        apiKeys:
          type: array
          items:
            type: string
        phoneNumbers:
          type: array
          items:
            type: string
        certificateFingerprint:
          type: ["string", "null"]
        certificateSubject:
          type: ["string", "null"]
        certificateIssuer:
          type: ["string", "null"]
        certificateNotBefore:
          type: ["string", "null"]
        certificateNotAfter:
          type: ["string", "null"]
        virusTotalScore:
          type: ["integer", "null"]
        virusTotalTotal:
          type: ["integer", "null"]
        abuseIpdbScore:
          type: ["integer", "null"]
        urlScanScore:
          type: ["integer", "null"]
        clusterId:
          type: ["string", "null"]
        anomalyScore:
          type: ["number", "null"]
        gnnMaliciousProb:
          type: ["number", "null"]
        pageRankScore:
          type: ["number", "null"]
      required:
        - sampleName
        - sha256
        - permissions
        - urls
        - domains
        - ipAddresses
        - apiKeys
        - phoneNumbers
    InvestigationSummary:
      type: object
      properties:
        id:
          type: integer
        sampleName:
          type: string
        sha256:
          type: string
        packageName:
          type: ["string", "null"]
        verdict:
          type: string
        riskLevel:
          type: string
        riskScore:
          type: integer
        confidence:
          type: string
        primaryThreatType:
          type: string
        clusterId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sampleName
        - sha256
        - verdict
        - riskLevel
        - riskScore
        - confidence
        - primaryThreatType
        - createdAt
    Investigation:
      type: object
      properties:
        id:
          type: integer
        sampleName:
          type: string
        sha256:
          type: string
        fuzzyHash:
          type: ["string", "null"]
        packageName:
          type: ["string", "null"]
        versionName:
          type: ["string", "null"]
        targetSdk:
          type: ["integer", "null"]
        compileSdk:
          type: ["integer", "null"]
        permissions:
          type: array
          items:
            type: string
        codeSnippets:
          type: ["string", "null"]
        urls:
          type: array
          items:
            type: string
        domains:
          type: array
          items:
            type: string
        ipAddresses:
          type: array
          items:
            type: string
        apiKeys:
          type: array
          items:
            type: string
        phoneNumbers:
          type: array
          items:
            type: string
        certificateFingerprint:
          type: ["string", "null"]
        certificateSubject:
          type: ["string", "null"]
        certificateIssuer:
          type: ["string", "null"]
        certificateNotBefore:
          type: ["string", "null"]
        certificateNotAfter:
          type: ["string", "null"]
        virusTotalScore:
          type: ["integer", "null"]
        virusTotalTotal:
          type: ["integer", "null"]
        abuseIpdbScore:
          type: ["integer", "null"]
        urlScanScore:
          type: ["integer", "null"]
        clusterId:
          type: ["string", "null"]
        anomalyScore:
          type: ["number", "null"]
        gnnMaliciousProb:
          type: ["number", "null"]
        pageRankScore:
          type: ["number", "null"]
        verdict:
          type: string
        riskLevel:
          type: string
        riskScore:
          type: integer
        confidence:
          type: string
        primaryThreatType:
          type: string
        analysis:
          $ref: "#/components/schemas/AnalysisReport"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sampleName
        - sha256
        - permissions
        - urls
        - domains
        - ipAddresses
        - apiKeys
        - phoneNumbers
        - verdict
        - riskLevel
        - riskScore
        - confidence
        - primaryThreatType
        - analysis
        - createdAt
    AnalysisReport:
      type: object
      properties:
        executiveSummary:
          type: string
        plainEnglishBrief:
          type: string
        behaviors:
          type: array
          items:
            $ref: "#/components/schemas/Behavior"
        codeFindings:
          type: array
          items:
            $ref: "#/components/schemas/CodeFinding"
        permissionAbuse:
          type: array
          items:
            $ref: "#/components/schemas/PermissionAbuse"
        networkInfrastructure:
          $ref: "#/components/schemas/NetworkInfrastructure"
        campaign:
          $ref: "#/components/schemas/CampaignAttribution"
        rootOffender:
          $ref: "#/components/schemas/RootOffender"
        riskMatrix:
          $ref: "#/components/schemas/RiskMatrix"
        prediction:
          $ref: "#/components/schemas/ThreatPrediction"
        mitreTactics:
          type: array
          items:
            $ref: "#/components/schemas/MitreTactic"
        iocs:
          type: array
          items:
            $ref: "#/components/schemas/Ioc"
        detectionRules:
          $ref: "#/components/schemas/DetectionRules"
        reasoningChain:
          type: array
          items:
            $ref: "#/components/schemas/ReasoningStep"
      required:
        - executiveSummary
        - plainEnglishBrief
        - behaviors
        - codeFindings
        - permissionAbuse
        - networkInfrastructure
        - campaign
        - rootOffender
        - riskMatrix
        - prediction
        - mitreTactics
        - iocs
        - detectionRules
        - reasoningChain
    Behavior:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        description:
          type: string
        evidence:
          type: array
          items:
            type: string
        confidence:
          type: integer
      required: [type, title, description, evidence, confidence]
    CodeFinding:
      type: object
      properties:
        title:
          type: string
        snippet:
          type: string
        meaning:
          type: string
        severity:
          type: string
      required: [title, snippet, meaning, severity]
    PermissionAbuse:
      type: object
      properties:
        permission:
          type: string
        riskLevel:
          type: string
        explanation:
          type: string
      required: [permission, riskLevel, explanation]
...
    RiskMatrix:
      type: object
      properties:
        severity:
          type: string
        impact:
          type: integer
        likelihood:
          type: integer
        composite:
          type: integer
        methodology:
          type: string
        cvssVector:
          type: ["string", "null"]
      required: [severity, impact, likelihood, composite, methodology]
...
    Ioc:
      type: object
      properties:
        type:
          type: string
        value:
          type: string
        context:
          type: string
      required: [type, value, context]
...
    CampaignDetail:
      type: object
      properties:
        clusterId:
          type: string
        clusterName:
          type: string
        sampleCount:
          type: integer
        topRiskLevel:
          type: string
        averageRiskScore:
          type: integer
        firstSeen:
          type: string
          format: date-time
        lastSeen:
          type: string
          format: date-time
        members:
          type: array
          items:
            $ref: "#/components/schemas/InvestigationSummary"
      required: [clusterId, clusterName, sampleCount, topRiskLevel, averageRiskScore, firstSeen, lastSeen, members]
...